# packages/api/src/openapi/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stagehand API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing security governance across organizational hierarchies",
    "contact": {
      "name": "Stagehand Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "CompanyHierarchy": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HierarchyItem"
          },
          {
            "type": "object",
            "properties": {
              "teams": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TeamHierarchy"
                }
              }
            }
          }
        ]
      },
      "HierarchyItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "organization",
              "company",
              "team",
              "project"
            ]
          },
          "isMember": {
            "type": "boolean",
            "description": "Whether the user is a direct member of this item"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrganizationHierarchy": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HierarchyItem"
          },
          {
            "type": "object",
            "properties": {
              "companies": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CompanyHierarchy"
                }
              }
            }
          }
        ]
      },
      "ProjectSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "repositoryUrl": {
            "type": "string"
          }
        }
      },
      "TeamHierarchy": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HierarchyItem"
          },
          {
            "type": "object",
            "properties": {
              "projects": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HierarchyItem"
                }
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/hierarchy": {
      "get": {
        "summary": "Get user's accessible organizational hierarchy",
        "description": "Returns the complete organizational hierarchy that the authenticated user has access to, structured as nested organizations, companies, teams, and projects",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Hierarchical structure of accessible organizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationHierarchy"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hierarchy/{resourceType}/{resourceId}/projects": {
      "get": {
        "summary": "Get all projects within a resource's hierarchy",
        "description": "Returns a flat list of all projects within the specified resource's hierarchy",
        "tags": [
          "Hierarchy"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "organization",
                "company",
                "team",
                "project"
              ]
            },
            "description": "Type of resource"
          },
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects within the resource hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to view this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ]
}